Revert "Move core functionality to a separate package. New nuget package version 6.3.0"
