--- Libraries/MvxExtensions.Core/MvxExtensions.Core.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="MSBuild.Sdk.Extras">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <AssemblyName>MvxExtensions.Core</AssemblyName>
-    <RootNamespace>MvxExtensions.Core</RootNamespace>
-    <PackageId>MvxExtensions.Core</PackageId>
-    <Description>
-      MvxExtensions provide a set of libraries that extend the awesome MvvmCross framework.
-      This package contains the 'Core' libraries of MvxExtensions with no dependencies to other MvxExtensions libraries.
-      This package is intended to provide some of the MvxExtensions functionalities without forcing to include other MvxExtensions packages
-      For more info check the documentation
-    </Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="MvvmCross" Version="6.4.2" />
-  </ItemGroup>
-  
-</Project>
