+++ Libraries/MvxExtensions.Forms/Views/ContentPage.cs
@@ -1,7 +1,6 @@
 ﻿using MvvmCross;
 using MvvmCross.Forms.Views;
 using MvvmCross.WeakSubscription;
-using MvxExtensions.Core.Extensions;
 using MvxExtensions.Extensions;
 using MvxExtensions.Plugins.Notification;
 using MvxExtensions.Plugins.Notification.Core;
