+++ Libraries/MvxExtensions/Extensions/NavigationExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using MvvmCross.Navigation;
 using MvvmCross.ViewModels;
-using MvxExtensions.Core.Extensions;
 using MvxExtensions.Statics;
 using System;
 using System.Threading;
