+++ Libraries/MvxExtensions/MvxExtensions.csproj
@@ -36,7 +36,6 @@ For more info check the documentation</Description>
   <ItemGroup>
     <ProjectReference Include="..\..\Plugins\Logger\MvxExtensions.Plugins.Logger.csproj" />
     <ProjectReference Include="..\..\Plugins\Notification\MvxExtensions.Plugins.Notification.csproj" />
-    <ProjectReference Include="..\MvxExtensions.Core\MvxExtensions.Core.csproj" />
   </ItemGroup>
 
 </Project>
