+++ MvxExtensions.sln
@@ -49,9 +49,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Playground.Forms.UI", "Samp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Playground.Forms.Core", "Samples\Playground\Playground.Forms\Playground.Forms.Core\Playground.Forms.Core.csproj", "{696B6AB9-CDDA-4D10-B4DF-117FBACC633A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MvxExtensions.Forms", "Libraries\MvxExtensions.Forms\MvxExtensions.Forms.csproj", "{CA8BBF3C-524F-4705-9E77-2CDD8FD84635}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MvxExtensions.Core", "Libraries\MvxExtensions.Core\MvxExtensions.Core.csproj", "{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MvxExtensions.Forms", "Libraries\MvxExtensions.Forms\MvxExtensions.Forms.csproj", "{CA8BBF3C-524F-4705-9E77-2CDD8FD84635}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -921,62 +919,6 @@ Global
 		{CA8BBF3C-524F-4705-9E77-2CDD8FD84635}.Release|x64.Build.0 = Release|Any CPU
 		{CA8BBF3C-524F-4705-9E77-2CDD8FD84635}.Release|x86.ActiveCfg = Release|Any CPU
 		{CA8BBF3C-524F-4705-9E77-2CDD8FD84635}.Release|x86.Build.0 = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|Any CPU.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|Any CPU.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|ARM.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|ARM.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|iPhone.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|iPhone.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|Mixed Platforms.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|x64.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|x64.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|x86.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Ad-Hoc|x86.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|Any CPU.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|Any CPU.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|ARM.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|ARM.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|iPhone.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|iPhone.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|Mixed Platforms.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|x64.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|x64.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|x86.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.AppStore|x86.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|ARM.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|x64.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Debug|x86.Build.0 = Debug|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|ARM.ActiveCfg = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|ARM.Build.0 = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|iPhone.Build.0 = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|x64.ActiveCfg = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|x64.Build.0 = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|x86.ActiveCfg = Release|Any CPU
-		{BEEB4537-05A9-4A91-B2E1-58214FB8EEC4}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
